Use `gemm`. Adds `cpu` feature. Removes matrixmultiply. Removes MKL support. Removes `cpu-seq-matmul`, `cpu-par-matmul`, `cpu-mkl-matmul` feature flags.
